Closes #13
Added the three .pngs for the falling stalactite
(fallingStalactite_bottom, fallingStalactite_top, and
fallingStalactite_whole) and imported them into unity. I tried
unsuccessfully applying them to the falling object Game Object. When I
looked at the object in the UI, I couldn't find the option to add it as
a material.